requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0 